requirements.txt
pandas
numpy
langdetect
sentence-transformers
faiss-cpu
pytest
langchain
langchain-community
langchain-openai
python-dotenv
pypdf
requests
